bithumb: move to separate file
